requirements.txt
streamlit
pandas
plotly
textblob
python-dotenv
requests
